Modify operation_over_instances and set_of_all to fix formatting issue
This incorporates the work from WW's branch 80-Fix-OperationOverInstances …
and updates that with substitutions of ExprTuple for ExprList and a minor bug
fix in set_of_all, all of which now allows the proper formatting for the subscripted domain output notation on a SetOfAll instance for when multiple
variables have the same domain or different domains. Brief testing shows this
is working well. Have not updated the corresponding parts of the 80-Fix
branch yet.
Related Issues: #80