Reorganize apply_commutation_thm() in generic_methods.py
with minor modifications to the initial permutation() method in the EnumSet
class to call, and feed permutation thms to, the apply_commutation_thm().
Will likely be changing the name of EnumSet permutation() method to allow
"permutation" to be used for a more general process in generic_methods.py
collection.
Related Issues #77, #69 .