Begin linking generic_permutation to TransRelUpdater()
This commits includes a new setEquivTransitivity thm in
set_theory/equivalence context and a new applyTransitivity() method in the
SetEquiv class (based on the analogous method in the Equals class). Still to
be dealt with is the fact that the TransRelUpdater() assumes we're dealing
with equalities instead of possible set equivalences (which is then causing a
headache elsewhere in the form of an infinite recursion inside the SetEquiv
class's applyTransitivity() method.