Augment Set.deduceEnumSubsetEq() method with Set argument
This will (eventually) allow users to provide an actual subset, in the form of an
enumerated set like Set(a, b, c), as the proposed subset instead of just a list
of indices to be used from the superset. This commit initially establishes this
as an alternative method Set.deduceEnumSubsetEqAlt() for testing, allowing
the subset argument and converting any subset to a list of indices. Still
incomplete. Related issues: #77 , #69 , #62 .