Update fields.py with FiniteField(n) class
Update fields.py with FiniteField(n) class to represent a finite field of size n, basically meaning the set {0, 1, 2, ..., n-1} with (implied) mod n addition and mod n multiplication associated with it. Technically for FiniteField(n) to be a field, n must be of the form p^{k}, where p is prime and k is a positive integer.